move verification folder into klayout
