+++ gplugins/klayout/get_netlist.py
