+++ gplugins/vlsir/tests/resources/pads_correct.cir
@@ -1,4 +1,4 @@
-; `circuit.Package` `gplugins.verification.example`
+; `circuit.Package` `gplugins.klayout.example`
 ; Generated by `vlsirtools.XyceNetlister`
 ;
