+++ gplugins/vlsir/tests/resources/pads_correct.scs
@@ -1,4 +1,4 @@
-// `circuit.Package` `gplugins.verification.example`
+// `circuit.Package` `gplugins.klayout.example`
 // Generated by `vlsirtools.SpectreNetlister`
 //
